+++ src/wrappers/vibe/__init__.py
@@ -1,0 +1,6 @@
+"""
+Vibe Coding Tool Wrapper Package
+"""
+from .vibe_coding_wrapper import vibe_coding
+
+__all__ = ['vibe_coding']